refactor(go): extract humanize into a shared module
The helper and its acronym tables were copy-pasted between
enums.ts and models.ts, so an acronym added to one copy would
silently drift the other. One module keeps enum and model doc
prose in sync.
