+++ src/go/models.ts
@@ -2,6 +2,7 @@ import type { Model, EmitterContext, GeneratedFile, TypeRef, Service } from '@wo
 import { walkTypeRef } from '@workos/oagen';
 import { mapTypeRef } from './type-map.js';
 import { className, domainFieldName } from './naming.js';
+import { humanize } from './humanize.js';
 import { lowerFirstForDoc, fieldDocComment, articleFor } from '../shared/naming-utils.js';
 import { isModelInScope, isScopedRun } from '../shared/resolved-ops.js';
 import { reconcileFlatBlocks, readPriorFile, parseFlatGoBlocks, type NamedBlock } from './flat-merge.js';
@@ -344,58 +345,6 @@ function crudEntityDomain(name: string): string | null {
     }
   }
   return null;
-}
-
-/** Known acronyms to preserve as single tokens during humanization. */
-const HUMANIZE_ACRONYMS: [RegExp, string][] = [
-  [/OAuth/g, 'OAUTH_ACRN'],
-  [/URN/g, 'URN_ACRN'],
-  [/IETF/g, 'IETF_ACRN'],
-  [/API/g, 'API_ACRN'],
-  [/SSO/g, 'SSO_ACRN'],
-  [/PKCE/g, 'PKCE_ACRN'],
-  [/JWT/g, 'JWT_ACRN'],
-  [/MFA/g, 'MFA_ACRN'],
-  [/TOTP/g, 'TOTP_ACRN'],
-  [/SAML/g, 'SAML_ACRN'],
-  [/SCIM/g, 'SCIM_ACRN'],
-  [/OIDC/g, 'OIDC_ACRN'],
-  [/CORS/g, 'CORS_ACRN'],
-  [/RBAC/g, 'RBAC_ACRN'],
-];
-
-const HUMANIZE_RESTORE: [RegExp, string][] = [
-  [/oauth_acrn/g, 'OAuth'],
-  [/urn_acrn/g, 'URN'],
-  [/ietf_acrn/g, 'IETF'],
-  [/api_acrn/g, 'API'],
-  [/sso_acrn/g, 'SSO'],
-  [/pkce_acrn/g, 'PKCE'],
-  [/jwt_acrn/g, 'JWT'],
-  [/mfa_acrn/g, 'MFA'],
-  [/totp_acrn/g, 'TOTP'],
-  [/saml_acrn/g, 'SAML'],
-  [/scim_acrn/g, 'SCIM'],
-  [/oidc_acrn/g, 'OIDC'],
-  [/cors_acrn/g, 'CORS'],
-  [/rbac_acrn/g, 'RBAC'],
-];
-
-function humanize(name: string): string {
-  // Replace known acronyms with placeholders before splitting
-  let s = name;
-  for (const [pattern, replacement] of HUMANIZE_ACRONYMS) {
-    s = s.replace(pattern, replacement);
-  }
-  // Split camelCase/PascalCase into words
-  let result = s.replace(/([a-z])([A-Z])/g, '$1 $2');
-  result = result.replace(/([A-Z]+)([A-Z][a-z])/g, '$1 $2');
-  result = result.toLowerCase();
-  // Restore acronyms
-  for (const [pattern, replacement] of HUMANIZE_RESTORE) {
-    result = result.replace(pattern, replacement);
-  }
-  return result;
 }
 
 function lowerFirst(s: string): string {
